Finished refactoring transforms
