+++ src/ImageSharp/Processing/Processors/Effects/OilPaintingProcessor{TPixel}.cs
@@ -1,5 +1,5 @@
 // Copyright (c) Six Labors and contributors.
-/// Licensed under the Apache License, Version 2.0.
+// Licensed under the Apache License, Version 2.0.
 
 using System;
 using System.Numerics;
