+++ src/ImageSharp/Processing/Processors/Transforms/CropProcessor.cs
@@ -1,32 +1,28 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using SixLabors.ImageSharp.Advanced;
-using SixLabors.ImageSharp.ParallelUtils;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing.Processors.Transforms
 {
     /// <summary>
-    /// Provides methods to allow the cropping of an image.
+    /// Defines a crop operation on an image.
     /// </summary>
-    /// <typeparam name="TPixel">The pixel format.</typeparam>
-    internal class CropProcessor<TPixel> : TransformProcessorBase<TPixel>
-        where TPixel : struct, IPixel<TPixel>
+    public sealed class CropProcessor : IImageProcessor
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="CropProcessor{TPixel}"/> class.
+        /// Initializes a new instance of the <see cref="CropProcessor"/> class.
         /// </summary>
         /// <param name="cropRectangle">The target cropped rectangle.</param>
         /// <param name="sourceSize">The source image size.</param>
         public CropProcessor(Rectangle cropRectangle, Size sourceSize)
         {
             // Check bounds here and throw if we are passed a rectangle exceeding our source bounds.
-            Guard.IsTrue(new Rectangle(Point.Empty, sourceSize).Contains(cropRectangle), nameof(cropRectangle), "Crop rectangle should be smaller than the source bounds.");
+            Guard.IsTrue(
+                new Rectangle(Point.Empty, sourceSize).Contains(cropRectangle),
+                nameof(cropRectangle),
+                "Crop rectangle should be smaller than the source bounds.");
             this.CropRectangle = cropRectangle;
         }
 
@@ -35,44 +31,11 @@ namespace SixLabors.ImageSharp.Processing.Processors.Transforms
         /// </summary>
         public Rectangle CropRectangle { get; }
 
-        /// <inheritdoc/>
-        protected override Image<TPixel> CreateDestination(Image<TPixel> source, Rectangle sourceRectangle)
+        /// <inheritdoc />
+        public IImageProcessor<TPixel> CreatePixelSpecificProcessor<TPixel>()
+            where TPixel : struct, IPixel<TPixel>
         {
-            // We will always be creating the clone even for mutate because we may need to resize the canvas
-            IEnumerable<ImageFrame<TPixel>> frames = source.Frames.Select(x => new ImageFrame<TPixel>(source.GetConfiguration(), this.CropRectangle.Width, this.CropRectangle.Height, x.Metadata.DeepClone()));
-
-            // Use the overload to prevent an extra frame being added
-            return new Image<TPixel>(source.GetConfiguration(), source.Metadata.DeepClone(), frames);
-        }
-
-        /// <inheritdoc/>
-        protected override void OnFrameApply(ImageFrame<TPixel> source, ImageFrame<TPixel> destination, Rectangle sourceRectangle, Configuration configuration)
-        {
-            // Handle resize dimensions identical to the original
-            if (source.Width == destination.Width && source.Height == destination.Height && sourceRectangle == this.CropRectangle)
-            {
-                // the cloned will be blank here copy all the pixel data over
-                source.GetPixelSpan().CopyTo(destination.GetPixelSpan());
-                return;
-            }
-
-            Rectangle rect = this.CropRectangle;
-
-            // Copying is cheap, we should process more pixels per task:
-            ParallelExecutionSettings parallelSettings = configuration.GetParallelSettings().MultiplyMinimumPixelsPerTask(4);
-
-            ParallelHelper.IterateRows(
-                rect,
-                parallelSettings,
-                rows =>
-                    {
-                        for (int y = rows.Min; y < rows.Max; y++)
-                        {
-                            Span<TPixel> sourceRow = source.GetPixelRowSpan(y).Slice(rect.Left);
-                            Span<TPixel> targetRow = destination.GetPixelRowSpan(y - rect.Top);
-                            sourceRow.Slice(0, rect.Width).CopyTo(targetRow);
-                        }
-                    });
+            return new CropProcessor<TPixel>(this);
         }
     }
 }
