+++ src/ImageSharp/Processing/Processors/Transforms/ProjectiveTransformProcessor{TPixel}.cs
@@ -1,0 +1,126 @@
+// Copyright (c) Six Labors and contributors.
+// Licensed under the Apache License, Version 2.0.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+
+using SixLabors.ImageSharp.Advanced;
+using SixLabors.ImageSharp.ParallelUtils;
+using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.Primitives;
+
+namespace SixLabors.ImageSharp.Processing.Processors.Transforms
+{
+    /// <summary>
+    /// Provides the base methods to perform non-affine transforms on an image.
+    /// </summary>
+    /// <typeparam name="TPixel">The pixel format.</typeparam>
+    internal class ProjectiveTransformProcessor<TPixel> : TransformProcessorBase<TPixel>
+        where TPixel : struct, IPixel<TPixel>
+    {
+        private readonly ProjectiveTransformProcessor definition;
+
+        public ProjectiveTransformProcessor(ProjectiveTransformProcessor definition)
+        {
+            this.definition = definition;
+        }
+
+        private Size TargetDimensions => this.definition.TargetDimensions;
+
+        /// <inheritdoc/>
+        protected override Image<TPixel> CreateDestination(Image<TPixel> source, Rectangle sourceRectangle)
+        {
+            // We will always be creating the clone even for mutate because we may need to resize the canvas
+            IEnumerable<ImageFrame<TPixel>> frames =
+                source.Frames.Select(x => new ImageFrame<TPixel>(source.GetConfiguration(), this.TargetDimensions.Width, this.TargetDimensions.Height, x.Metadata.DeepClone()));
+
+            // Use the overload to prevent an extra frame being added
+            return new Image<TPixel>(source.GetConfiguration(), source.Metadata.DeepClone(), frames);
+        }
+
+        /// <inheritdoc/>
+        protected override void OnFrameApply(ImageFrame<TPixel> source, ImageFrame<TPixel> destination, Rectangle sourceRectangle, Configuration configuration)
+        {
+            Matrix4x4 transformMatrix = this.definition.TransformMatrix;
+
+            // Handle tranforms that result in output identical to the original.
+            if (transformMatrix.Equals(default) || transformMatrix.Equals(Matrix4x4.Identity))
+            {
+                // The clone will be blank here copy all the pixel data over
+                source.GetPixelSpan().CopyTo(destination.GetPixelSpan());
+                return;
+            }
+
+            int width = this.TargetDimensions.Width;
+            var targetBounds = new Rectangle(Point.Empty, this.TargetDimensions);
+
+            // Convert from screen to world space.
+            Matrix4x4.Invert(transformMatrix, out Matrix4x4 matrix);
+
+            IResampler sampler = this.definition.Sampler;
+
+            if (sampler is NearestNeighborResampler)
+            {
+                ParallelHelper.IterateRows(
+                    targetBounds,
+                    configuration,
+                    rows =>
+                        {
+                            for (int y = rows.Min; y < rows.Max; y++)
+                            {
+                                Span<TPixel> destRow = destination.GetPixelRowSpan(y);
+
+                                for (int x = 0; x < width; x++)
+                                {
+                                    Vector2 point = TransformUtils.ProjectiveTransform2D(x, y, matrix);
+                                    int px = (int)MathF.Round(point.X);
+                                    int py = (int)MathF.Round(point.Y);
+
+                                    if (sourceRectangle.Contains(px, py))
+                                    {
+                                        destRow[x] = source[px, py];
+                                    }
+                                }
+                            }
+                        });
+
+                return;
+            }
+
+            var kernel = new TransformKernelMap(configuration, source.Size(), destination.Size(), sampler);
+            try
+            {
+                ParallelHelper.IterateRowsWithTempBuffer<Vector4>(
+                    targetBounds,
+                    configuration,
+                    (rows, vectorBuffer) =>
+                        {
+                            Span<Vector4> vectorSpan = vectorBuffer.Span;
+                            for (int y = rows.Min; y < rows.Max; y++)
+                            {
+                                Span<TPixel> targetRowSpan = destination.GetPixelRowSpan(y);
+                                PixelOperations<TPixel>.Instance.ToVector4(configuration, targetRowSpan, vectorSpan);
+                                ref float ySpanRef = ref kernel.GetYStartReference(y);
+                                ref float xSpanRef = ref kernel.GetXStartReference(y);
+
+                                for (int x = 0; x < width; x++)
+                                {
+                                    // Use the single precision position to calculate correct bounding pixels
+                                    // otherwise we get rogue pixels outside of the bounds.
+                                    Vector2 point = TransformUtils.ProjectiveTransform2D(x, y, matrix);
+                                    kernel.Convolve(point, x, ref ySpanRef, ref xSpanRef, source.PixelBuffer, vectorSpan);
+                                }
+
+                                PixelOperations<TPixel>.Instance.FromVector4Destructive(configuration, vectorSpan, targetRowSpan);
+                            }
+                        });
+            }
+            finally
+            {
+                kernel.Dispose();
+            }
+        }
+    }
+}
