+++ src/ImageSharp/Processing/Processors/Transforms/SkewProcessor.cs
@@ -9,14 +9,12 @@ using SixLabors.Primitives;
 namespace SixLabors.ImageSharp.Processing.Processors.Transforms
 {
     /// <summary>
-    /// Provides methods that allow the skewing of images.
+    /// Defines a skew transformation applicable to an <see cref="Image"/>.
     /// </summary>
-    /// <typeparam name="TPixel">The pixel format.</typeparam>
-    internal class SkewProcessor<TPixel> : AffineTransformProcessor<TPixel>
-        where TPixel : struct, IPixel<TPixel>
+    internal class SkewProcessor : AffineTransformProcessor
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="SkewProcessor{TPixel}"/> class.
+        /// Initializes a new instance of the <see cref="SkewProcessor"/> class.
         /// </summary>
         /// <param name="degreesX">The angle in degrees to perform the skew along the x-axis.</param>
         /// <param name="degreesY">The angle in degrees to perform the skew along the y-axis.</param>
@@ -27,7 +25,7 @@ namespace SixLabors.ImageSharp.Processing.Processors.Transforms
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="SkewProcessor{TPixel}"/> class.
+        /// Initializes a new instance of the <see cref="SkewProcessor"/> class.
         /// </summary>
         /// <param name="degreesX">The angle in degrees to perform the skew along the x-axis.</param>
         /// <param name="degreesY">The angle in degrees to perform the skew along the y-axis.</param>
