+++ src/ImageSharp/Processing/TransformExtensions.cs
@@ -3,60 +3,53 @@
 
 using System.Numerics;
 
-using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing.Processors.Transforms;
 using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing
 {
     /// <summary>
-    /// Adds extensions that allow the application of composable transform operations to the <see cref="Image{TPixel}"/> type.
+    /// Adds extensions that allow the application of composable transform operations to the <see cref="Image"/> type.
     /// </summary>
     public static class TransformExtensions
     {
         /// <summary>
         /// Performs an affine transform of an image.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="source">The image to transform.</param>
         /// <param name="builder">The affine transform builder.</param>
         /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> source,
-            AffineTransformBuilder builder)
-            where TPixel : struct, IPixel<TPixel>
-            => Transform(source, builder, KnownResamplers.Bicubic);
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext source,
+            AffineTransformBuilder builder) =>
+            Transform(source, builder, KnownResamplers.Bicubic);
 
         /// <summary>
         /// Performs an affine transform of an image using the specified sampling algorithm.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
-        /// <param name="ctx">The <see cref="IImageProcessingContext{TPixel}"/>.</param>
+        /// <param name="ctx">The <see cref="IImageProcessingContext"/>.</param>
         /// <param name="builder">The affine transform builder.</param>
         /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> ctx,
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext ctx,
             AffineTransformBuilder builder,
-            IResampler sampler)
-            where TPixel : struct, IPixel<TPixel>
-            => ctx.Transform(new Rectangle(Point.Empty, ctx.GetCurrentSize()), builder, sampler);
+            IResampler sampler) =>
+            ctx.Transform(new Rectangle(Point.Empty, ctx.GetCurrentSize()), builder, sampler);
 
         /// <summary>
         /// Performs an affine transform of an image using the specified sampling algorithm.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
-        /// <param name="ctx">The <see cref="IImageProcessingContext{TPixel}"/>.</param>
+        /// <param name="ctx">The <see cref="IImageProcessingContext"/>.</param>
         /// <param name="sourceRectangle">The source rectangle</param>
         /// <param name="builder">The affine transform builder.</param>
         /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> ctx,
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext ctx,
             Rectangle sourceRectangle,
             AffineTransformBuilder builder,
             IResampler sampler)
-            where TPixel : struct, IPixel<TPixel>
         {
             Matrix3x2 transform = builder.BuildMatrix(sourceRectangle);
             Size targetDimensions = TransformUtils.GetTransformedSize(sourceRectangle.Size, transform);
@@ -66,69 +59,61 @@ namespace SixLabors.ImageSharp.Processing
         /// <summary>
         /// Performs an affine transform of an image using the specified sampling algorithm.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
-        /// <param name="ctx">The <see cref="IImageProcessingContext{TPixel}"/>.</param>
+        /// <param name="ctx">The <see cref="IImageProcessingContext"/>.</param>
         /// <param name="sourceRectangle">The source rectangle</param>
         /// <param name="transform">The transformation matrix.</param>
         /// <param name="targetDimensions">The size of the result image.</param>
         /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> ctx,
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext ctx,
             Rectangle sourceRectangle,
             Matrix3x2 transform,
             Size targetDimensions,
             IResampler sampler)
-            where TPixel : struct, IPixel<TPixel>
         {
             return ctx.ApplyProcessor(
-                new AffineTransformProcessor<TPixel>(transform, sampler, targetDimensions),
+                new AffineTransformProcessor(transform, sampler, targetDimensions),
                 sourceRectangle);
         }
 
         /// <summary>
         /// Performs a projective transform of an image.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="source">The image to transform.</param>
         /// <param name="builder">The affine transform builder.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> source,
-            ProjectiveTransformBuilder builder)
-            where TPixel : struct, IPixel<TPixel>
-            => Transform(source, builder, KnownResamplers.Bicubic);
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext source,
+            ProjectiveTransformBuilder builder) =>
+            Transform(source, builder, KnownResamplers.Bicubic);
 
         /// <summary>
         /// Performs a projective transform of an image using the specified sampling algorithm.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
-        /// <param name="ctx">The <see cref="IImageProcessingContext{TPixel}"/>.</param>
+        /// <param name="ctx">The <see cref="IImageProcessingContext"/>.</param>
         /// <param name="builder">The projective transform builder.</param>
         /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> ctx,
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext ctx,
             ProjectiveTransformBuilder builder,
-            IResampler sampler)
-            where TPixel : struct, IPixel<TPixel>
-            => ctx.Transform(new Rectangle(Point.Empty, ctx.GetCurrentSize()), builder, sampler);
+            IResampler sampler) =>
+            ctx.Transform(new Rectangle(Point.Empty, ctx.GetCurrentSize()), builder, sampler);
 
         /// <summary>
         /// Performs a projective transform of an image using the specified sampling algorithm.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
-        /// <param name="ctx">The <see cref="IImageProcessingContext{TPixel}"/>.</param>
+        /// <param name="ctx">The <see cref="IImageProcessingContext"/>.</param>
         /// <param name="sourceRectangle">The source rectangle</param>
         /// <param name="builder">The projective transform builder.</param>
         /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> ctx,
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext ctx,
             Rectangle sourceRectangle,
             ProjectiveTransformBuilder builder,
             IResampler sampler)
-            where TPixel : struct, IPixel<TPixel>
         {
             Matrix4x4 transform = builder.BuildMatrix(sourceRectangle);
             Size targetDimensions = TransformUtils.GetTransformedSize(sourceRectangle.Size, transform);
@@ -138,23 +123,21 @@ namespace SixLabors.ImageSharp.Processing
         /// <summary>
         /// Performs a projective transform of an image using the specified sampling algorithm.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
-        /// <param name="ctx">The <see cref="IImageProcessingContext{TPixel}"/>.</param>
+        /// <param name="ctx">The <see cref="IImageProcessingContext"/>.</param>
         /// <param name="sourceRectangle">The source rectangle</param>
         /// <param name="transform">The transformation matrix.</param>
         /// <param name="targetDimensions">The size of the result image.</param>
         /// <param name="sampler">The <see cref="IResampler"/> to perform the resampling.</param>
-        /// <returns>The <see cref="Image{TPixel}"/></returns>
-        public static IImageProcessingContext<TPixel> Transform<TPixel>(
-            this IImageProcessingContext<TPixel> ctx,
+        /// <returns>The <see cref="Image{TPixel}"/>.</returns>
+        public static IImageProcessingContext Transform(
+            this IImageProcessingContext ctx,
             Rectangle sourceRectangle,
             Matrix4x4 transform,
             Size targetDimensions,
             IResampler sampler)
-            where TPixel : struct, IPixel<TPixel>
         {
             return ctx.ApplyProcessor(
-                new ProjectiveTransformProcessor<TPixel>(transform, sampler, targetDimensions),
+                new ProjectiveTransformProcessor(transform, sampler, targetDimensions),
                 sourceRectangle);
         }
     }
