+++ tests/ImageSharp.Tests/Processing/Transforms/PadTest.cs
@@ -1,15 +1,13 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Processing;
 using SixLabors.ImageSharp.Processing.Processors.Transforms;
+
 using Xunit;
 
 namespace SixLabors.ImageSharp.Tests.Processing.Transforms
 {
-    using SixLabors.ImageSharp.Processing;
-    using SixLabors.ImageSharp.Processing.Processors.Transforms;
-
     public class PadTest : BaseImageOperationsExtensionTest
     {
         [Fact]
@@ -20,7 +18,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Transforms
             IResampler sampler = KnownResamplers.NearestNeighbor;
 
             this.operations.Pad(width, height);
-            ResizeProcessor<Rgba32> resizeProcessor = this.Verify<ResizeProcessor<Rgba32>>();
+            ResizeProcessor resizeProcessor = this.Verify<ResizeProcessor>();
 
             Assert.Equal(width, resizeProcessor.Width);
             Assert.Equal(height, resizeProcessor.Height);
