+++ tests/ImageSharp.Tests/Processing/Transforms/SkewTest.cs
@@ -16,7 +16,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Transforms
         public void SkewXYCreateSkewProcessorWithAnglesSet()
         {
             this.operations.Skew(10, 20);
-            SkewProcessor<Rgba32> processor = this.Verify<SkewProcessor<Rgba32>>();
+            SkewProcessor processor = this.Verify<SkewProcessor>();
 
             Assert.Equal(10, processor.DegreesX);
             Assert.Equal(20, processor.DegreesY);
